chore: refactor letter handling
Removes code duplication and uses a letter queue instead of task
notifications which simplifies the code.
