+++ apps/battery-monitor/src/freertos-tasks.c
@@ -7,15 +7,13 @@
 #include "battery.h"
 #include "ck-data.h"
 #include "ck-rx-letters.h"
-#include "led.h"
 #include "peripherals.h"
 #include "ports.h"
 #include "potentiometer.h"
 
 // STM32Common
 #include "error.h"
-#include "postmaster-hal.h"
-#include "spi-flash.h"
+#include "letter-reader.h"
 
 // CK
 #include "mayor.h"
@@ -42,36 +40,31 @@ static TaskHandle_t battery_report_task;
 static StaticTask_t battery_report_buf;
 static StackType_t battery_report_stack[configMINIMAL_STACK_SIZE];
 
-// CAN Kingdom process received letters task
-static TaskHandle_t process_letter_task;
-static StaticTask_t process_letter_buf;
-// Need at least one page of stack for interacting with the SPI flash.
-static StackType_t
-    process_letter_stack[SPI_FLASH_PAGE_SIZE + configMINIMAL_STACK_SIZE];
-
 void battery_monitor(void *unused);
 void battery_report(void *unused);
-void process_letter(void *unused);
 void battery_monitor_timer(TimerHandle_t timer);
 void battery_report_timer(TimerHandle_t timer);
 void update_pages(void);
 void send_docs(void);
-void dispatch_letter(ck_letter_t *letter);
 int handle_letter(const ck_folder_t *folder, const ck_letter_t *letter);
 
 void task_init(void) {
-  battery_monitor_task = xTaskCreateStatic(
-      battery_monitor, "battery monitor", configMINIMAL_STACK_SIZE, NULL,
-      LOWEST_TASK_PRIORITY + 1, battery_monitor_stack, &battery_monitor_buf);
-
   battery_report_task = xTaskCreateStatic(
       battery_report, "battery report", configMINIMAL_STACK_SIZE, NULL,
       LOWEST_TASK_PRIORITY, battery_report_stack, &battery_report_buf);
 
-  process_letter_task = xTaskCreateStatic(
-      process_letter, "process letter",
-      SPI_FLASH_PAGE_SIZE + configMINIMAL_STACK_SIZE, NULL,
-      LOWEST_TASK_PRIORITY + 2, process_letter_stack, &process_letter_buf);
+  battery_monitor_task = xTaskCreateStatic(
+      battery_monitor, "battery monitor", configMINIMAL_STACK_SIZE, NULL,
+      LOWEST_TASK_PRIORITY + 1, battery_monitor_stack, &battery_monitor_buf);
+
+  letter_reader_cfg_t letter_reader_cfg = {
+      .priority = LOWEST_TASK_PRIORITY + 2,
+      .app_letter_handler_func = handle_letter,
+  };
+
+  if (init_letter_reader_task(letter_reader_cfg) != APP_OK) {
+    error();
+  }
 }
 
 void set_task_periods(task_periods_t *periods) {
@@ -149,35 +142,6 @@ void battery_report(void *unused) {
   }
 }
 
-void process_letter(void *unused) {
-  (void)unused;
-
-  peripherals_t *peripherals = get_peripherals();
-
-  CAN_RxHeaderTypeDef header;
-  uint8_t data[CK_CAN_MAX_DLC];
-  ck_letter_t letter;
-
-  for (;;) {
-    if (HAL_CAN_ActivateNotification(&peripherals->common_peripherals->hcan,
-                                     CAN_IT_RX_FIFO0_MSG_PENDING) != HAL_OK) {
-      printf("Error activating interrupt.\r\n");
-      error();
-    }
-    ulTaskNotifyTake(pdTRUE, portMAX_DELAY);
-
-    // Process all messages
-    while (HAL_CAN_GetRxMessage(&peripherals->common_peripherals->hcan,
-                                CAN_RX_FIFO0, &header, data) == HAL_OK) {
-      if (ck_correct_letter_received() != CK_OK) {
-        printf("CAN Kingdom error in ck_correct_letter_received().\r\n");
-      }
-      letter = frame_to_letter(&header, data);
-      dispatch_letter(&letter);
-    }
-  }
-}
-
 void battery_monitor_timer(TimerHandle_t timer) {
   (void)timer;
   xTaskNotifyGive(battery_monitor_task);
@@ -222,29 +186,6 @@ void send_docs(void) {
   }
 }
 
-void dispatch_letter(ck_letter_t *letter) {
-  ck_folder_t *folder = NULL;
-
-  // Check for default letter
-  if (ck_is_default_letter(letter) == CK_OK) {
-    if (ck_default_letter_received() != CK_OK) {
-      printf("CAN Kingdom error in ck_default_letter_received().\r\n");
-    }
-  }
-  // Check for king's letter
-  else if (ck_is_kings_envelope(&letter->envelope) == CK_OK) {
-    if (ck_process_kings_letter(letter) != CK_OK) {
-      printf("failed to process king's letter.\r\n");
-    }
-  }
-  // Check for any other letter
-  else if (ck_get_envelopes_folder(&letter->envelope, &folder) == CK_OK) {
-    if (handle_letter(folder, letter) != APP_OK) {
-      printf("failed to process page.\r\n");
-    }
-  }
-}
-
 int handle_letter(const ck_folder_t *folder, const ck_letter_t *letter) {
   ck_data_t *ck_data = get_ck_data();
 
@@ -270,15 +211,5 @@ void HAL_ADC_ConvCpltCallback(ADC_HandleTypeDef *hadc) {
   (void)hadc;
   BaseType_t higher_priority_task_woken = pdFALSE;
   vTaskNotifyGiveFromISR(battery_monitor_task, &higher_priority_task_woken);
-  portYIELD_FROM_ISR(higher_priority_task_woken);
-}
-
-// Deactivate interrupt, then signal task. Let the task reactivate the
-// interrupt.
-void HAL_CAN_RxFifo0MsgPendingCallback(CAN_HandleTypeDef *hcan) {
-  HAL_CAN_DeactivateNotification(hcan, CAN_IT_RX_FIFO0_MSG_PENDING);
-
-  BaseType_t higher_priority_task_woken = pdFALSE;
-  vTaskNotifyGiveFromISR(process_letter_task, &higher_priority_task_woken);
   portYIELD_FROM_ISR(higher_priority_task_woken);
 }
