+++ apps/servo/src/freertos-tasks.c
@@ -10,9 +10,8 @@
 
 // STM32Common
 #include "error.h"
-#include "postmaster-hal.h"
+#include "letter-reader.h"
 #include "rover.h"
-#include "spi-flash.h"
 
 // CK
 #include "king.h"
@@ -45,13 +44,6 @@ static TaskHandle_t report_task;
 static StaticTask_t report_buf;
 static StackType_t report_stack[configMINIMAL_STACK_SIZE];
 
-// CAN Kingdom process received letters task
-static TaskHandle_t process_letter_task;
-static StaticTask_t process_letter_buf;
-// Need at least one page of stack for interacting with the SPI flash.
-static StackType_t
-    process_letter_stack[SPI_FLASH_PAGE_SIZE + configMINIMAL_STACK_SIZE];
-
 void king(void *unused);
 // King helpers
 void king_timer_callback(TimerHandle_t timer);
@@ -62,12 +54,10 @@ void start_communication(void);
 
 void measure(void *unused);
 void report(void *unused);
-void process_letter(void *unused);
 void measure_timer(TimerHandle_t timer);
 void report_timer(TimerHandle_t timer);
 
 void send_docs(void);
-void dispatch_letter(ck_letter_t *letter);
 int handle_letter(const ck_folder_t *folder, const ck_letter_t *letter);
 
 void task_init(void) {
@@ -79,10 +69,14 @@ void task_init(void) {
       xTaskCreateStatic(measure, "measure", configMINIMAL_STACK_SIZE, NULL,
                         LOWEST_TASK_PRIORITY + 1, measure_stack, &measure_buf);
 
-  process_letter_task = xTaskCreateStatic(
-      process_letter, "process letter",
-      SPI_FLASH_PAGE_SIZE + configMINIMAL_STACK_SIZE, NULL,
-      LOWEST_TASK_PRIORITY + 2, process_letter_stack, &process_letter_buf);
+  letter_reader_cfg_t letter_reader_cfg = {
+      .priority = LOWEST_TASK_PRIORITY + 2,
+      .app_letter_handler_func = handle_letter,
+  };
+
+  if (init_letter_reader_task(letter_reader_cfg) != APP_OK) {
+    error();
+  }
 
   king_task =
       xTaskCreateStatic(king, "king", configMINIMAL_STACK_SIZE, NULL,
@@ -416,35 +410,6 @@ void report(void *unused) {
   }
 }
 
-void process_letter(void *unused) {
-  (void)unused;
-
-  peripherals_t *peripherals = get_peripherals();
-
-  CAN_RxHeaderTypeDef header;
-  uint8_t data[CK_CAN_MAX_DLC];
-  ck_letter_t letter;
-
-  for (;;) {
-    if (HAL_CAN_ActivateNotification(&peripherals->common_peripherals->hcan,
-                                     CAN_IT_RX_FIFO0_MSG_PENDING) != HAL_OK) {
-      printf("Error activating interrupt.\r\n");
-      error();
-    }
-    ulTaskNotifyTake(pdTRUE, portMAX_DELAY);
-
-    // Process all messages
-    while (HAL_CAN_GetRxMessage(&peripherals->common_peripherals->hcan,
-                                CAN_RX_FIFO0, &header, data) == HAL_OK) {
-      if (ck_correct_letter_received() != CK_OK) {
-        printf("CAN Kingdom error in ck_correct_letter_received().\r\n");
-      }
-      letter = frame_to_letter(&header, data);
-      dispatch_letter(&letter);
-    }
-  }
-}
-
 void measure_timer(TimerHandle_t timer) {
   (void)timer;
   xTaskNotifyGive(measure_task);
@@ -472,29 +437,6 @@ void send_docs(void) {
   }
   if (ck_send_document(ck_data->h_bridge_current_folder->folder_no) != CK_OK) {
     printf("failed to send doc.\r\n");
-  }
-}
-
-void dispatch_letter(ck_letter_t *letter) {
-  ck_folder_t *folder = NULL;
-
-  // Check for default letter
-  if (ck_is_default_letter(letter) == CK_OK) {
-    if (ck_default_letter_received() != CK_OK) {
-      printf("CAN Kingdom error in ck_default_letter_received().\r\n");
-    }
-  }
-  // Check for king's letter
-  else if (ck_is_kings_envelope(&letter->envelope) == CK_OK) {
-    if (ck_process_kings_letter(letter) != CK_OK) {
-      printf("failed to process king's letter.\r\n");
-    }
-  }
-  // Check for any other letter
-  else if (ck_get_envelopes_folder(&letter->envelope, &folder) == CK_OK) {
-    if (handle_letter(folder, letter) != APP_OK) {
-      printf("failed to process page.\r\n");
-    }
   }
 }
 
@@ -533,14 +475,4 @@ void HAL_ADC_ConvCpltCallback(ADC_HandleTypeDef *hadc) {
     vTaskNotifyGiveFromISR(measure_task, &higher_priority_task_woken);
     portYIELD_FROM_ISR(higher_priority_task_woken);
   }
-}
-
-// Deactivate interrupt, then signal task. Let the task reactivate the
-// interrupt.
-void HAL_CAN_RxFifo0MsgPendingCallback(CAN_HandleTypeDef *hcan) {
-  HAL_CAN_DeactivateNotification(hcan, CAN_IT_RX_FIFO0_MSG_PENDING);
-
-  BaseType_t higher_priority_task_woken = pdFALSE;
-  vTaskNotifyGiveFromISR(process_letter_task, &higher_priority_task_woken);
-  portYIELD_FROM_ISR(higher_priority_task_woken);
 }
