+++ apps/servo/src/main.c
@@ -12,7 +12,6 @@
 
 // CK
 #include "mayor.h"
-#include "postmaster-hal.h"
 
 // STM32Common
 #include "clock.h"
